+++ src/design-system/theme/index.ts
@@ -1,2 +1,3 @@
+export * from './density'
 export * from './types'
 export * from './validation'